feat: Fix and improve cli
* Add a confirmation message before deletion
* Fix insertion
* Add the possibility to specify date range and start for raw measure
insert

+++ dacc/cli/cli.py
@@ -7,6 +7,9 @@ from dacc.fixtures import fixtures
 def reset_tables():
     """Reset all tables in database"""
 
+    click.confirm(
+        "This will remove ALL DATA in database. Are you sure?", abort=True
+    )
     print("Reset all tables in database...")
     db.drop_all()
     db.create_all()
@@ -26,12 +29,17 @@ def show_table(table_name):
 def insert_fixtures_definition():
     """Insert fixtures in database"""
 
-    fixtures.insert_fixtures_measures_definition()
+    fixtures.insert_measures_definition_from_file()
+    db.session.commit()
 
 
 @dacc.cli.command("insert-random-measures")
-@click.option("-n", default=1, show_default=True)
-def insert_fixtures(n):
+@click.option("-n", "n_measures", default=1, show_default=True)
+@click.option("-d", "--days", default=1, show_default=True)
+@click.option(
+    "--starting-day", "starting_day", default="2021-05-01", show_default=True
+)
+def insert_fixtures(n_measures, days, starting_day):
     """Insert random measures in database"""
 
-    fixtures.insert_random_raw_measures(n_measures)
+    fixtures.insert_random_raw_measures(n_measures, days, starting_day)
